backend/requirements.txt
fastapi
uvicorn[standard]
APScheduler
pydantic
motor
python-dotenv 